qlog: change event time field from f32 to f64
This change is a stepping stop to supporting
https://datatracker.ietf.org/doc/html/draft-ietf-quic-qlog-main-schema-13.
It increases the precision of the qlog event time field from
32-bit to 64-bit floating point.

While it would have been nice to do this as one atomic change,
it ends up touching a lot of dependent crates for such a small
technical change. Hence, spliting it out to make other qlog
update PRs easier to review.

Changes:
- Updated qlog Event and JsonEvent time fields from f32 to f64
- Updated Eventable trait set_time() method signature
- Updated Event::with_time() and with_time_ex() constructors
- Updated all time conversions using as_secs_f32() to as_secs_f64()
- Updated qlog-dancer types to use f64 for time coordinates
- RTT values remain as f32 since we aren't travelling to outerspace yet.

+++ qlog-dancer/src/plots/minmax.rs
@@ -26,12 +26,12 @@
 
 #[derive(Clone, Copy)]
 pub struct XMinMax {
-    pub min: f32,
-    pub max: f32,
+    pub min: f64,
+    pub max: f64,
 }
 
 impl XMinMax {
-    pub fn new(min: f32, max: f32, start: Option<f32>, end: Option<f32>) -> Self {
+    pub fn new(min: f64, max: f64, start: Option<f64>, end: Option<f64>) -> Self {
         let mut minmax = Self { min, max };
 
         if let Some(s) = start {
@@ -45,7 +45,7 @@ impl XMinMax {
         minmax
     }
 
-    pub fn range(&self) -> std::ops::Range<f32> {
+    pub fn range(&self) -> std::ops::Range<f64> {
         self.min..self.max
     }
 }
@@ -57,8 +57,8 @@ pub struct XYMinMax<Y> {
 
 impl<Y> XYMinMax<Y> {
     pub fn init(
-        x_data_range: std::ops::Range<f32>, x_start: Option<f32>,
-        x_end: Option<f32>, y_range: std::ops::Range<Y>,
+        x_data_range: std::ops::Range<f64>, x_start: Option<f64>,
+        x_end: Option<f64>, y_range: std::ops::Range<Y>,
     ) -> Self {
         let x =
             XMinMax::new(x_data_range.start, x_data_range.end, x_start, x_end);
